+++ app/src/main/java/ru/artempugachev/concoord/DmsToDmFragment.java
@@ -199,9 +199,9 @@ public class DmsToDmFragment extends Fragment {
             if(MainActivity.checkDMCoord(latDeg, latMin, -90, 90)) {
                 if (MainActivity.checkDMCoord(lonDeg, lonMin, -180, 180)) {
                     Coordinate lat = new Coordinate(Integer.parseInt(latDeg),
-                            Integer.parseInt(latMin), latLabel);
+                            Double.parseDouble(latMin), latLabel);
                     Coordinate lon = new Coordinate(Integer.parseInt(lonDeg),
-                            Integer.parseInt(lonMin), lonLabel);
+                            Double.parseDouble(lonMin), lonLabel);
                     updateDmsFields(lat, lon);
 
                 } else {
